Clarify how capable dragonslayer adversary is (#15)
Small wording tweak; saying the adversary is not "generally capable" seems a bit off given Tony struggled to win against it (Tony may not be the strongest Go player but I believe he is a general intelligence). So I've replaced it with saying adversary is not stronger than a Go amateur, which seems more precise.