game_info: Improved step count matching (#118)
Issue: KataGoVisualizer doesn't parse the step count for victims with names like "vit-s650m". This is because the parsing regex always expects a `-` after `-s<number>`. This leads to display issues in the training progress visualizer since it separates victims by looking at their visit counts.

This PR fixes this issue and adds a comment explaining the regex.